git-debimport heimdal_1.6~git20120403+dfsg1.orig.tar.gz

+++ lib/asn1/symbol.h
@@ -36,6 +36,7 @@
 #ifndef _SYMBOL_H
 #define _SYMBOL_H
 
+#include <roken.h>
 #include "asn1_queue.h"
 
 enum typetype {
@@ -78,7 +79,7 @@ struct value {
     } type;
     union {
 	int booleanvalue;
-	int integervalue;
+	int64_t integervalue;
 	char *stringvalue;
 	struct objid *objectidentifiervalue;
     } u;
@@ -109,8 +110,12 @@ struct tagtype {
 };
 
 struct range {
-    int min;
-    int max;
+    /*
+     * We can't represent unsigned 64-bit ranges because max might be
+     * negative...
+     */
+    int64_t min;
+    int64_t max;
 };
 
 enum ctype { CT_CONTENTS, CT_USER } ;
